@@ -12,45 +12,54 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.7",
-    "@angular/cdk": "^6.4.3",
-    "@angular/common": "6.0.7",
-    "@angular/compiler": "6.0.7",
-    "@angular/core": "6.0.7",
-    "@angular/forms": "6.0.7",
-    "@angular/http": "6.0.7",
-    "@angular/material": "^6.4.3",
-    "@angular/platform-browser": "6.0.7",
-    "@angular/platform-browser-dynamic": "6.0.7",
-    "@angular/router": "6.0.7",
-    "@ngrx/effects": "6.0.1",
-    "@ngrx/store": "6.0.1",
-    "@ngrx/store-devtools": "6.0.1",
-    "hoek": "5.0.3",
-    "core-js": "2.5.4",
-    "rxjs": "6.2.1",
-    "zone.js": "^0.8.19"
+    "@angular/animations": "7.0.2",
+    "@angular/cdk": "7.3.0",
+    "@angular/common": "7.0.2",
+    "@angular/compiler": "7.0.2",
+    "@angular/core": "7.0.2",
+    "@angular/forms": "7.0.2",
+    "@angular/http": "7.0.2",
+    "@angular/material": "7.3.0",
+    "@angular/platform-browser": "7.0.2",
+    "@angular/platform-browser-dynamic": "7.0.2",
+    "@angular/router": "7.0.2",
+    "@ngrx/effects": "6.1.2",
+    "@ngrx/router-store": "6.1.2",
+    "@ngrx/store": "6.1.2",
+    "@ngrx/store-devtools": "6.1.2",
+    "core-js": "2.5.7",
+    "hammerjs": "^2.0.8",
+    "rxjs": "6.3.3",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "6.0.8",
-    "@angular/compiler-cli": "6.0.7",
-    "@angular/language-service": "6.0.7",
-    "@types/jasmine": "2.8.8",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "6.0.60",
-    "codelyzer": "4.2.1",
-    "jasmine-core": "3.1.0",
+    "@angular-devkit/build-angular": "^0.13.8",
+    "@angular-devkit/build-ng-packagr": "^0.13.8",
+    "@angular/cli": "7.0.3",
+    "@angular/compiler-cli": "^8.1.3",
+    "@angular/language-service": "7.0.2",
+    "@types/jasmine": "2.8.9",
+    "@types/jasminewd2": "2.0.5",
+    "@types/lodash": "4.14.117",
+    "@types/moment": "2.13.0",
+    "@types/node": "10.12.1",
+    "@types/numeral": "0.0.25",
+    "@types/selectables": "^1.4.0",
+    "@types/signalr": "^2.2.35",
+    "codelyzer": "4.5.0",
+    "glob": "^7.1.3",
+    "jasmine-core": "2.99.1",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "^3.0.0",
+    "karma": "^4.2.0",
     "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-coverage-istanbul-reporter": "2.0.4",
     "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "1.2.0",
-    "protractor": "^5.4.0",
-    "ts-mockito": "2.3.0",
-    "ts-node": "5.0.1",
-    "tslint": "5.9.1",
-    "typescript": "2.7.2"
+    "karma-jasmine-html-reporter": "0.2.2",
+    "karma-junit-reporter": "^1.2.0",
+    "protractor": "5.4.1",
+    "svgstore": "^3.0.0-2",
+    "ts-node": "7.0.1",
+    "tslint": "^5.11.0",
+    "typescript": "3.4.5"
   }
 }
