@@ -1,10 +1,10 @@
 import { Injectable } from '@angular/core';
 import { Actions, Effect, ofType } from '@ngrx/effects';
-import { map, catchError, mergeMap } from 'rxjs/operators';
-import { Observable } from 'rxjs';
+import { switchMap } from 'rxjs/operators';
 
-import * as CardListActions from './cardlist.actions';
-import { CardListService, CardListMockService} from '@cv/CardList/service';
+import * as CardListActions from '@cv/CardList/store/cardlist.actions';
+import { CardListService } from '@cv/CardList/service';
+import { CardItem } from '@cv/CardList/models';
 
 @Injectable()
 export class CardListEffects {
@@ -16,20 +16,8 @@ export class CardListEffects {
 
   @Effect() getItems$ = this.action$.pipe(
     ofType(CardListActions.GetCards.TYPE),
-    mergeMap((action: CardListActions.GetCards) =>
-      this.service.getItems(action.itemIds).pipe(
-        map(items => new CardListActions.GetCardsSuccess(items))
-      )
+    switchMap((action: CardListActions.GetCards) => 
+      this.service.getItems(action.itemIds).pipe(switchMap((items: CardItem[]) => [new CardListActions.GetCardsSuccess(items)]))
     )
   );
-
-  // @Effect() addItem$ = this.action$.pipe(
-  //   ofType(CardListActions.AddCard.TYPE),
-  //   mergeMap((action: CardListActions.AddCard) =>
-  //     this.service.addItem(action.item)
-  //     .pipe(
-  //       map(item => new CardListActions.AddCardSuccess(item))
-  //     )
-  //   )
-  // )
 }
