@@ -1,4 +1,4 @@
-export * from './cardsearch.module';
+export * from './deckBuilder.module';
 export * from './components';
 export * from './models';
 export * from './service';
