@@ -4,7 +4,6 @@ using QueueHandler.Queues;
 using RabbitMQ.Client;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-
 using OfficeOpenXml;
 using System.IO;
 using System;
@@ -21,6 +20,7 @@ namespace LocalData.Service
 		#endregion
 
 		#region -  Constructor  -
+
 		public LocalDataWorker(ConnectionFactory factory) : base(factory)
         {
 			ReadAndReply<LoadCardDataRequest, (CollectionData[], LocalCardData[])>(Queueing.Queues.GetCollectionData, true, LoadDataFromSpreadsheet);
@@ -30,46 +30,28 @@ namespace LocalData.Service
 
 		#region -  Event Handlers  -
 
-		private Task<(CollectionData[], LocalCardData[])> LoadDataFromSpreadsheet(object sender, ReceiveEventArgs<LoadCardDataRequest> e)
+		private async Task<(CollectionData[], LocalCardData[])> LoadDataFromSpreadsheet(object sender, ReceiveEventArgs<LoadCardDataRequest> e)
 		{
-			//Interrogate the spreadsheet and return the name/quantity pairs there
-			var heldCards = GetLocalCardData(sender, e).Result;
-			//Get the latest scryfall info that we've downloaded
-			var scryfallData = GetScryfallData(scryfallDataLocation).Result;
-
 			List<CollectionData> collectionData = new List<CollectionData>();
-            List<LocalCardData> errorCards = new List<LocalCardData>();
-            heldCards.ForEach(card =>
+			List<LocalCardData> errorCards = new List<LocalCardData>();
+			List<LocalCardData> cardsInSpreadsheet = await MapSpreadsheetData(spreadsheetFileLocation);
+			List<LocalScryfallData> scryfallData = GetScryfallData(scryfallDataLocation).Result;
+
+			cardsInSpreadsheet.ForEach(card =>
 			{
-				try
-				{
-					//Try and find our held card in the scryfall data
-                    var foundName = scryfallData.First(d => d.Name == card.Name && d.Set.ToUpper() == card.Set.ToUpper());
-                    if(foundName != null)
-                    {
-						//if we have a match, extract the data
-                        collectionData.Add(ModelMapper.MapCollectionData(foundName, card));
-                    } else
-                    {
-                        errorCards.Add(card);
-                    }
-				}
-				catch
-				{
-                    errorCards.Add(card);
+				//Try and find our held card in the scryfall data
+				LocalScryfallData foundName = scryfallData.FirstOrDefault(d => d.Name == card.Name && d.Set.ToUpper() == card.Set.ToUpper());
+                if(foundName != null)
+                {
+					collectionData.Add(ModelMapper.MapCollectionData(foundName, card)); //if we have a match, extract the data
+				} else
+                {
+					errorCards.Add(card); //If we don't have a match, add the card to our investigation list
                 }
 			});
 
-            var result = (collectionData.ToArray(), errorCards.ToArray());
-			return Task.FromResult(result);
+			return (collectionData.ToArray(), errorCards.ToArray());
 		}
-
-		private async Task<List<LocalCardData>> GetLocalCardData(object sender, ReceiveEventArgs<LoadCardDataRequest> e)
-        {
-            List<LocalCardData> allCards = new List<LocalCardData>();
-            allCards.AddRange(await MapSpreadsheetData(spreadsheetFileLocation));
-            return allCards;
-        }
 
         private Task<List<LocalScryfallData>> GetScryfallData(string scryfallDataLocation)
         {
@@ -93,26 +75,23 @@ namespace LocalData.Service
 
         private Task<List<LocalCardData>> MapSpreadsheetData(string fileLocation)
         {
-            FileInfo file = new FileInfo(fileLocation);
-            List<LocalCardData> spreadsheetCardList = new List<LocalCardData>();
-            var package = new ExcelPackage(file);
-            var totalSheets = package.Workbook.Worksheets.Count;
+            List<LocalCardData> spreadsheetCardList = new List<LocalCardData>(); //Empty list to hold spreadsheet data
+			ExcelPackage package = new ExcelPackage(new FileInfo(fileLocation)); //Get the spreadshet info
 
-			//For each sheet
-            for(int i = 1; i <= totalSheets; i++)
+            for(int i = 1; i <= package.Workbook.Worksheets.Count; i++)  //For each sheet in the workbook
             {
                 ExcelWorksheet workSheet = package.Workbook.Worksheets[i];
-				//for each row in the sheet
-                for (int row = workSheet.Dimension.Start.Row + 1; row <= workSheet.Dimension.End.Row; row++)
-                {
+				
+                for (int row = workSheet.Dimension.Start.Row + 1; row <= workSheet.Dimension.End.Row; row++) //For each row in the sheet
+				{
                     try
                     {
-						//as long as we have a card name, add the name and quantity to our list
+						//as long as we have a card name, add the name, quantity and set to our list
                         if (workSheet.Cells[row, 2].Value.ToString() != null)
                         {
-                            string spreadsheetCardName = workSheet.Cells[row, 2].Value.ToString();
-                            int cardQuantity = Int32.Parse(workSheet.Cells[row, 3].Value.ToString());
 							string set = workSheet.Cells[row, 1].Value.ToString();
+							string spreadsheetCardName = workSheet.Cells[row, 2].Value.ToString();
+                            int cardQuantity = Int32.Parse(workSheet.Cells[row, 3].Value.ToString());
 							spreadsheetCardList.Add(new LocalCardData { Name = spreadsheetCardName, Quantity = cardQuantity, Set = set });
                         }
                     } catch (Exception ex) {
